Move TextDeserializer to a dedicated file and document it
