+++ src/de/map.rs
@@ -4,7 +4,8 @@ use crate::{
     de::key::QNameDeserializer,
     de::resolver::EntityResolver,
     de::simple_type::SimpleTypeDeserializer,
-    de::{str2bool, DeEvent, Deserializer, Text, XmlRead, TEXT_KEY, VALUE_KEY},
+    de::text::TextDeserializer,
+    de::{str2bool, DeEvent, Deserializer, XmlRead, TEXT_KEY, VALUE_KEY},
     encoding::Decoder,
     errors::serialize::DeError,
     events::attributes::IterState,
@@ -1012,158 +1013,6 @@ where
             // SAFETY: the other events are filtered in `variant_seed()`
             _ => unreachable!("Only `Start` or `Text` events are possible here"),
         }
-    }
-
-    fn newtype_variant_seed<T>(self, seed: T) -> Result<T::Value, Self::Error>
-    where
-        T: DeserializeSeed<'de>,
-    {
-        seed.deserialize(self)
-    }
-
-    #[inline]
-    fn tuple_variant<V>(self, len: usize, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        self.deserialize_tuple(len, visitor)
-    }
-
-    #[inline]
-    fn struct_variant<V>(
-        self,
-        fields: &'static [&'static str],
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        self.deserialize_struct("", fields, visitor)
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-struct TextDeserializer<'de>(Text<'de>);
-
-impl<'de> TextDeserializer<'de> {
-    /// Returns a next string as concatenated content of consequent [`Text`] and
-    /// [`CData`] events, used inside [`deserialize_primitives!()`].
-    ///
-    /// [`Text`]: crate::events::Event::Text
-    /// [`CData`]: crate::events::Event::CData
-    #[inline]
-    fn read_string(self) -> Result<Cow<'de, str>, DeError> {
-        Ok(self.0.text)
-    }
-}
-
-impl<'de> de::Deserializer<'de> for TextDeserializer<'de> {
-    type Error = DeError;
-
-    deserialize_primitives!();
-
-    fn deserialize_unit<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_unit()
-    }
-
-    fn deserialize_option<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        if self.0.is_empty() {
-            visitor.visit_none()
-        } else {
-            visitor.visit_some(self)
-        }
-    }
-
-    /// Forwards deserialization of the inner type. Always calls [`Visitor::visit_newtype_struct`]
-    /// with this deserializer.
-    fn deserialize_newtype_struct<V>(
-        self,
-        _name: &'static str,
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_newtype_struct(self)
-    }
-
-    /// This method deserializes a sequence inside of element that itself is a
-    /// sequence element:
-    ///
-    /// ```xml
-    /// <>
-    ///   ...
-    ///   inner sequence as xs:list
-    ///   ...
-    /// </>
-    /// ```
-    fn deserialize_seq<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        SimpleTypeDeserializer::from_text_content(self.0).deserialize_seq(visitor)
-    }
-
-    #[inline]
-    fn deserialize_struct<V>(
-        self,
-        _name: &'static str,
-        _fields: &'static [&'static str],
-        _visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        Err(DeError::ExpectedStart)
-    }
-
-    fn deserialize_enum<V>(
-        self,
-        _name: &'static str,
-        _variants: &'static [&'static str],
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_enum(self)
-    }
-
-    #[inline]
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        self.deserialize_str(visitor)
-    }
-}
-
-impl<'de> de::EnumAccess<'de> for TextDeserializer<'de> {
-    type Error = DeError;
-    type Variant = Self;
-
-    fn variant_seed<V>(self, seed: V) -> Result<(V::Value, Self::Variant), Self::Error>
-    where
-        V: DeserializeSeed<'de>,
-    {
-        let name = seed.deserialize(BorrowedStrDeserializer::<DeError>::new(TEXT_KEY))?;
-        Ok((name, self))
-    }
-}
-
-impl<'de> de::VariantAccess<'de> for TextDeserializer<'de> {
-    type Error = DeError;
-
-    #[inline]
-    fn unit_variant(self) -> Result<(), Self::Error> {
-        Ok(())
     }
 
     fn newtype_variant_seed<T>(self, seed: T) -> Result<T::Value, Self::Error>
